requirements.txt
Brotli==1.0.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dash==0.21.1
dash-core-components==0.23.0
dash-html-components==0.11.0
dash-renderer==0.13.0
Flask==1.1.2
Flask-Compress==1.5.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
langdetect==1.0.8
MarkupSafe==1.1.1
nltk==3.5
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.3
pkg-resources==0.0.0
plotly==4.7.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.7
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
textblob==0.15.1
tqdm==4.46.0
tweepy==3.6.0
Unidecode==1.0.22
urllib3==1.25.9
Werkzeug==1.0.1
